chore: intermediate rename-step to achieve changing casing of  filenames
Explanation: Git is not good at changing casing of file names, so we'll first rewrite the name, then get back to the correct name with the proper casing
